refactor(StorageService): add descriptive file name
